# apps/backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Plataforma SaaS especializada em cafeicultura para gestão operacional, produtiva, financeira e analítica de propriedades cafeeiras.",
        "title": "CafeOS API",
        "contact": {},
        "license": {
            "name": "Proprietary"
        },
        "version": "0.1.0"
    },
    "host": "localhost:5001",
    "basePath": "/api/v1/{tenant_id}",
    "paths": {
...
        "/api/v1/{tenant_id}/agricultural-products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista todos os produtos agrícolas pertencentes ao tenant",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agricultural-products (Produtos Agrícolas)"
                ],
                "summary": "Listar produtos agrícolas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.AgriculturalProduct"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/{tenant_id}/stock/items": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista todos os itens de estoque pertencentes ao tenant",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock (Estoque)"
                ],
                "summary": "Listar itens de estoque",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.StockItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra um novo item de estoque no tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock (Estoque)"
                ],
                "summary": "Criar item de estoque",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do item de estoque",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createStockItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.StockItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/{tenant_id}/stock/items/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna um único item de estoque",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock (Estoque)"
                ],
                "summary": "Obter item de estoque por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do Item de Estoque",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.StockItem"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza dados do item de estoque por ID (atualização parcial - somente os campos informados são alterados)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock (Estoque)"
                ],
                "summary": "Atualizar item de estoque",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do Item de Estoque",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados do item de estoque",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.StockItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.StockItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Exclui um item de estoque por ID",
                "tags": [
                    "stock (Estoque)"
                ],
                "summary": "Excluir item de estoque",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do Item de Estoque",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/{tenant_id}/stock/movements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista todas as movimentações de estoque pertencentes ao tenant",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock (Estoque)"
                ],
                "summary": "Listar movimentações de estoque",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.StockMovement"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra uma movimentação de estoque (entrada/saída) para um item de estoque",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock (Estoque)"
                ],
                "summary": "Registrar movimentação de estoque",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados da movimentação de estoque",
                        "name": "movement",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.recordMovementRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.StockMovement"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.AgriculturalProduct": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/entity.ProductType"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "entity.ProductType": {
            "type": "string",
            "enum": [
                "fertilizante",
                "defensivo",
                "combustivel",
                "outro"
            ],
            "x-enum-varnames": [
                "ProdFertilizante",
                "ProdDefensivo",
                "ProdCombustivel",
                "ProdOutro"
            ]
        },
...
        "entity.StockItem": {
            "type": "object",
            "properties": {
                "batch": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expiry_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "min_stock": {
                    "type": "number"
                },
                "notes": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                },
                "tenant_id": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entity.StockMovement": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "item_id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                },
                "reference": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "type": {
                    "description": "\"in\" | \"out\"",
                    "type": "string"
                }
            }
        },
...
        "handler.createStockItemRequest": {
            "type": "object",
            "properties": {
                "batch": {
                    "type": "string"
                },
                "expiry_date": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "min_stock": {
                    "type": "number"
                },
                "notes": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "handler.recordMovementRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "item_id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                },
                "reference": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token. Use: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}